requirements.txt
absl-py==1.0.0
aiofiles==22.1.0
aiosqlite==0.19.0
amqp==5.1.1
ansi2html==1.8.0
anyio==3.6.2
argcomplete==1.10.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
bleach==6.0.0
braintree==4.26.0
cachetools==5.0.0
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
channels==4.0.0
channels-redis==4.0.0
chardet==3.0.4
charset-normalizer==2.0.10
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cn2an==0.5.22
colorama==0.4.5
comm==0.1.2
compressed-rtf==1.0.6
cryptography==36.0.1
cx-Oracle==8.3.0
cycler==0.11.0
dash==2.8.1
dash-bootstrap-components==1.4.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.6
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
diff-match-patch==20200713
dj-static==0.0.6
Django==3.2.7
django-admin-import==0.3.0
django-autoslug==1.9.9
django-celery-results==2.4.0
django-crispy-forms==1.14.0
django-import-export==2.8.0
django-plotly-dash==2.1.3
django-taggit==0.22.2
django-thread==0.0.1
djangorestframework==3.13.1
docx==0.2.4
docx2txt==0.8
dpd-components==0.1.0
ebcdic==1.1.1
EbookLib==0.17.1
entrypoints==0.4
ephem==4.1.4
et-xmlfile==1.1.0
executing==1.2.0
extract-msg==0.28.7
fastjsonschema==2.16.3
Flask==2.2.3
flatbuffers==2.0
flower==1.2.0
fonttools==4.32.0
fqdn==1.5.1
gast==0.5.3
google-auth==2.18.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.44.0
gunicorn==20.1.0
h5py==3.6.0
humanize==4.4.0
idna==3.3
IMAPClient==2.1.0
import-export==0.2.67.dev6
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-dash==0.4.2
jupyter-events==0.6.3
jupyter-ydoc==0.2.4
jupyter_client==8.2.0
jupyter_core==5.2.0
jupyter_server==2.5.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.22.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
kombu==5.2.4
libclang==13.0.0
lxml==4.7.1
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==2.1.2
martor==1.6.13
matplotlib==3.5.1
matplotlib-inline==0.1.6
mistune==2.0.5
msal==1.24.1
msgpack==1.0.4
myBasics==1.0.0
myHttp==1.1.4
mySecrets==1.0.4
nbclassic==1.0.0
nbclient==0.7.3
nbconvert==7.3.0
nbformat==5.8.0
nest-asyncio==1.5.6
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.22.1
oauthlib==3.1.1
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.9
opt-einsum==3.3.0
outlook==1.0.5
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pdfminer.six==20191110
pdfrw==0.4
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==3.0.0
plotly==5.13.1
proces==0.1.7
prometheus-client==0.14.1
prompt-toolkit==1.0.18
protobuf==3.20.0
psutil==5.9.4
pure-eval==0.2.2
pvlib==0.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.12.0
pyephem==9.99
Pygments==2.14.0
PyJWT==2.3.0
pyparsing==3.0.8
PyPDF4==1.27.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
python-json-logger==2.0.7
python-pptx==0.6.21
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
redis==4.5.1
regex==2023.12.25
requests==2.29.0
requests-oauthlib==1.3.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roseta==0.1.9
rosetta==0.3
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.2
simplegeneric==0.8.1
six==1.12.0
sklearn==0.0
sniffio==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
SQLAlchemy==2.0.12
sqlparse==0.4.2
stack-data==0.6.2
static3==0.7.0
tablib==3.2.1
tenacity==8.2.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.17.1
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
timedelta==2020.12.3
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.1
tqdm==4.66.2
traitlets==5.9.0
typing_extensions==4.2.0
tzdata==2021.5
tzlocal==4.1
uri-template==1.2.0
urllib3==1.26.8
vine==5.0.0
waitress==2.1.2
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
whitenoise==6.0.0
wrapt==1.14.0
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.0
y-py==0.5.9
ypy-websocket==0.8.2
